vqe/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
qiskit>=1.0.0
qiskit-aer>=0.13.0
scipy>=1.11.0
numpy>=1.24.0
anthropic>=0.25.0
python-dotenv>=1.0.0
